初期コミット

+++ app.py
@@ -35,7 +35,8 @@ gg_tran_coin = int(arrival_coin * TAN93)
 # 取引手数料（7%）
 ta_tran = int(arrival_coin * TAN07)
 # GGドル
-gg_doll = int(gg_tran_coin / exchange_rate)
+gg_doll = gg_tran_coin / exchange_rate
+# result_gg_doll = 
 
 if st.button("計算"):
     st.success(f"GGドル交換は  {gg_tran_coin:,} 円")
